refactor code and fix imports errors
